minsylls and maxsylls bug fix
Adding values to minsylls and maxsylls in the rule file didn't change the syllable number output. 
I erased the lines that fixed the number of syllables (18, 19). 
I also changed all the others expressions of minsylls and maxsylls into their standard (rule) expressions : minSyllables, minsyllables and maxSyllables, maxsyllables into minsylls and maxsylls.
This makes it necessary to have entered the values for minsylls = ... and maxsylls = ... in the -PARAMS in the rule file, alongside with -CATEGORIES and -SYLLABLES. 
I'm only a novice, so the changes may not be correct, but the issue seems to be fixed with that.